Store search param with zustand
