+++ frontend/src/app/@modal/(.)product/[productId]/page.tsx
@@ -2,7 +2,8 @@
 
 import ProductModal from '@/components/product/product-modal';
 import { useRouter } from 'next/navigation';
-import { use } from 'react';
+import { use, useEffect } from 'react';
+import useSearchParamsStore from '@/stores/useSearchParamsStore';
 
 type PageParams = {
   productId: string;
@@ -12,6 +13,14 @@ const ProductModalPage = ({ params }: { params: Promise<PageParams> }) => {
   const router = useRouter();
   const resolvedParams = use(params);
   const { productId } = resolvedParams;
+  const { preservedSearchParams } = useSearchParamsStore();
+
+  useEffect(() => {
+    if (preservedSearchParams && preservedSearchParams.toString()) {
+      const newUrl = `${window.location.pathname}?${preservedSearchParams.toString()}`;
+      window.history.replaceState(null, '', newUrl);
+    }
+  }, [preservedSearchParams]);
 
   const handleClose = () => {
     router.back();
