+++ frontend/src/app/page.tsx
@@ -6,12 +6,12 @@ import ProductList from '@/components/product/product-list';
 import { useSearchParams } from 'next/navigation';
 import { useDebounce } from 'use-debounce';
 import { useEffect } from 'react';
-import { usePreservedSearchParams } from './layout';
+import useSearchParamsStore from '@/stores/useSearchParamsStore';
 
 const Home = () => {
   const pageTitle = 'Home - DINCT';
   const searchParams = useSearchParams();
-  const { preservedSearchParams } = usePreservedSearchParams();
+  const { preservedSearchParams } = useSearchParamsStore();
   // Use current search params or preserved search params
   const currentSearchParams = searchParams.toString() ? searchParams : preservedSearchParams;
   const searchQuery = currentSearchParams?.get('search') || '';
